[groovy] add GrCallTypeCalculator
The API breaks dependency of custom call calculation onto method call expression. This allows to provide IntelliJ
with return type for some call outside of any context, e.g. in the end it doesn't matter if the method was invoked with
method call expression or index access or getter method was invoked via property access.

- port existing implementations of GrCallExpressionTypeCalculator, finally schedule it for removal;
- put callable property logic outside of call type api;
- extract closure methods into separate implementation.