Fixing CPP-14615 - Show more detailed clang errors
Method HighlightInfo.encodeTooltip() used to have unclear behaviour depending on whether description was
found inside tooltip or not.

The tooltip was processed two times: when creating HiglightInfo and when getting tooltip from it.
When creating:
1) In case description was inside tooltip, encodeTooltip() unescaped tooltip, replaced description with placeholders
and returned the resulting string. Thus the stored tooltip was something strange: if client provided a valid html,
then it might become invalid because escaped sequences were unescaped.
2) In case description was not inside tooltip, it remained unchanged. Thus, if client provided a valid html, then it
remained valid.

When getting tooltip:
In method HighlightInfo.getTooltip() the stored tooltip was used as following:
1) If it was encoded (it has placeholders inside), then the escaped description was inserted inside the stored tooltip,
wrapped in html tags and returned. Obviously, the result might be invalid html, because the stored tooltip was
unescaped when creating the HighlightInfo!
2) If it wasn't encoded, the tooltip was used as is. Thus, it must be a valid html.

So, the problem was in unescaping a valid tooltip. Probably, it was done to search for a description (description
supposed to be a plain text, thus not escaped). Seems, that the solution is to search for escaped description
instead of unescaping tooltip.