Use `toString` as a sort key for PsiDirectoryNode (GO-6466)
If an IDE supports attaching project, the custom sorting key is used for directories
in order to lift the main project up in the Project View: empty string for the main project and `getTitle` for other directories.

The problem is that getTitle returns an entire path to a directory, so sorting
is performed by path, not by name. It's not an intuitive behavior for the attached projects.
And it's completely confusing for other directory nodes that do not contain
absolute path in their presentation (see GO-6466).

To address this, let's sort directories by toString() value, as it is also used
while default sorting. toString() value contains only the name of a directory.