Refresh the Git root candidate when searching for .git inside it.
Otherwise if a Git directory is created just before searching for it, the VFS may be not aware of it as it happens in IDEA-169020.

Introduce a utility method refreshAndFindChild() which could be better in some conditions than refreshAndFindFileByPath/ByIoFile.

(cherry picked from commit 5341f4288b1d89d4954e2e35021f7772b1ebb1dd)