IDEA-202796 IDEA-198143 git: fix hanging remote operations on Ubuntu
We use SSH_ASKPASS to handle passphrase requests (and other interactive input) in "Native SSH" mode.
openssh ignores SSH_ASKPASS if there are TTY attached to the process.

This can happen if IDE is executed from terminal. In this case, passphrase prompt
will be shown in the terminal, while git process will seem "hanging", silently awaiting for user input.

Some Linux setups can pass TTY[1-7] from Xorg to the IDE even if it's started with desktop launcher.
This is known to happen on Ubuntu that uses gdm3 by default (while lightdm is not affected).
In this case passphrase is asked in system TTY and is not visible to the user.

We fix the issue by detaching git process from IDE terminal session,
forcing openssh to use our SSH_ASKPASS handler.