github: let disable on-site Merge actions for a project
There are projects (e.g. intellij-community itself) the GitHub remote of which is not the primary remote, but is just a mirror, which is synchronized from time to time from the main repository hosted elsewhere. In these cases it makes no sense and is confusing to allow merge directly through GitHub API.

Such projects could manually define the following property in .idea/vcs.xml, and Merge actions will be disabled.

<component name="GithubSharedSettings">
  <option name="MERGE_ACTIONS_ALLOWED" value="false" />
</component>