git: fetch all remotes when called from Git | Fetch action; clarify and cleanup the fetch API and implementation
It fixes IDEA-201193, and as a side effect fetches multiple remotes in parallel.

There is no need in complicated ErrorRepoResult and RepoResult hierarchy.
There is no need to mention a repository in the progress indicator title, since the fetch is parallel.
Each FetchTask is fetching a single remote and thus returns a SingleRemoteResult. After all FetchTasks complete, their results are collected into a list of RepoResults - each per repository.