update trove4j library to the version which is published to Maven Central
There are no changes in the code between 1.0.20160824 and 1.0.20181211. The latter just includes a javadoc artifact and has detailed descriptions in pom.xml. The update is needed just for consistency with Kotlin compiler, which uses modules from IntelliJ platform and need to take trove4j from Maven Central (see IDEA-203634).