[file-history] track paths more precisely during history calculation
File history worked incorrectly when a file was renamed and later on a new file with the same name appeared. Commits before the first rename appeared in history for a new file, while they should not have, since they are a part of different history.

In order to fix the problem, paths tracking was rewritten. It now takes deleted flag into the account, and if file is deleted, tries to find a conflicting rename.