Select first leaf that contains an error on Build view after build
When a build is performed, the build view displays a tree containg the
tasks including their status. If there are errors or warning it can be
difficult to find them between the rest of the nodes. To help with that
this change selects the first node that has an error (or warning if no
errors are found) and for wich none of its descendants has an error.

Fixes https://issuetracker.google.com/issues/76453233