PY-32773: Move code that closes all opened branches to java side.
==Python part:
Test runners do not report suites, only tests with qualified names.
To create tree, our code splits test names, and we also must explicitly close each branch.
This was done on Python side using atexit, but several frameworks launch atexit explicitly. so we decided to move it to Java side.
This change is also required to support multiprocess.

==Platform side:
We added special callback that listener may implement to be called before onTestingFinished. We need it because onTestingFinished makrs unfinished tests as skipped, and we need to close them my our code.