git: refresh only .git/refs/tags when a tag is created or deleted
There is no need to refresh the whole .git directory, it is enough to refresh .git/refs/tags and .git/packed-refs (if the tag was packed).

GitRepositoryFiles#refresh() is not needed anymore, thus remove it. Rename refreshNonTrackedData to a more specific refreshTagsFiles(), since it is used for tags only.

Fix for IDEA-204248