[vcs-log] properly support working with filter pairs in filter models
Some filter models work with several filters simultaneously, for example FileFilterModel sets both structure and root filters, and TextFilterModel works with text and hash filters. For FileFilterModel a special composite filter class, VcsLogFileFilter, was used. It implemented VcsLogFilter interface, but was not actually a filter, which was confusing. On the other hand, TextFilterModel did not have a special filter class and just secretly created a hash filter. In order to unify working with such models, FilterPair class is introduced, and a descendant of FilterModel for working with filter pairs is extracted. Both FileFilterModel and TextFilterModel are adjusted to work with FilterPair.