[JCEF/markdown] clear focus on client size when browser gains it
Currently, when I click into the browser - it gains it, but at the same time java focus remains unchanged on the client side (for instance, the caret continues blinking in the editor). Then, when I click back somewhere in the client side (e.g. in the editor) - native focus remains in the browser.

The fix clears focus that allows for triggering java/native focus request in the client when native focus is in the browser and I click in the client.