fixup! Move SSH executable switch to registry
IDEA provides 2 SSH executables to be used with git - Builtin and Native.
The only reason to use Builtin instead of Native was inability of the IDE to handle interactive prompts from Native SSH.
In all other aspects, Native is better than the builtin - it supports more features, security protocols and keys.

The issue with interactive prompts was fixed in d6b37824, and now it feels like Builtin executable is not needed anymore,
and there is no need to confuse users with the option in settings.

Keeping the option in registry as a fallback for some not foreseen cases.