vcs: don't search for .git roots on module change events (IDEA-201588)
rootsChanged() is fired too often, not only on manual changes of the project model, e.g. it is fired during compilation, and scanning the whole tree is expensive.

At the same time, it doesn't look like a valuable feature: the case is rare and there are workarounds (add the mapping manually, or reload the project).

Thus simply removing can on every module event.