IDEA-204831 ensure that on close project actual state is saved
It is not expected and not good that application saved on project close and dispose — it is caller responsibility to call it if need.

In our case, RecentProjectsManager MUST save new state AFTER project closed and disposed (because ProjectManager.openedProjects list is checked). So, in any case we MUST call app.saveSettings() after `projectFrameClosed` event