Don't modify TestPlan
Although `TestPlan` exposes an `add(TestIdentifier)` method, it was only
ever intended for internal use by JUnit's `DefaultLauncher` which
already adds dynamic test identifiers to the `TestPlan`. The JUnit team
is going to make the `add()` throw an exception on modifications from
listeners. Thus, this commit removes the unnecessary call.