[groovy] gdsl: handle methods and variables separately
This change allows to skip processing of methods or variables or both when its not needed.

At this point org.jetbrains.plugins.groovy.lang.resolve.processors.KindsResolverProcessor
emits a warning when element passed into it has unexpected kind.
This will cause an error when all such cases are identified and fixed.