git: fix NPE in case of submodules in detached HEAD with enabled synchoronous roots control
And also simplify the logic: in case of sync-control fail if there is any repository in the detached HEAD state, otherwise fail only if all repositories are in the detached HEAD state. Detached HEAD submodules are ignored from this logic, because they are updated via a separate procedure.

Follow-up: 719a321c