git: correctly detect if there are local changes that should be saved before update: IDEA-76760
The "AD" status (staged addition + unstaged deletion of the same file) is not detected as a local change, thus we add an additional check for `git diff --staged`.

Note: stashing "AD" would stash only the addition, thus after stash-unstash process the file will be created back. However, it seems to be a less serious issue than unability to update.