Enable FS case sensitivity by default for ASWB on Mac
Although the filesystem used by Macs is case-insensitive, all our
users' code lives in google3, which is not. There are several places
in google3 where two files have the same case-insensitive name, which
causes errors in the VFS because IntelliJ defaults to treating the
file system as case-insensitive for Macs.

This CL resolves the issue by enabling the idea.case.sensitive.fs
property by default for ASWB on Mac release builds.

Bug: 123586448
Test: test_studio.py
Change-Id: I7b6ac108cc6bfe7610a6c8c1f863fbd80b4da132