git: don't save the answer to any generic input
SSH can ask literally about anything and in any format, and the same message can be used for different questions. Better not to risk here, and wait for an explicit request to support a specific message (like it is done by askConfirmationInput).

Follow-up: 5f2c5a4e
Relates to IDEA-206134
IDEA-CR-42499