git: make sure the known answer for an ssh confirmation is used only once per root
Otherwise it could be possible that the answer was incorrect, or it was a second answer, in which cases we'd better ask user explicitly without using the answer saved for another root.

This fix uses the fact that the GitNativeSshGuiAuthenticator is created per git command, thus per root, as opposed to the GitAuthenticationGate which is created per multi-root stack of commands.

Follow-up: 5f2c5a4e
Relates to IDEA-206134
IDEA-CR-42499