[vcs-log] show only changes that affect selected file paths in compare actions if the corresponding option is selected
When folder history is displayed in the Log, comparing actions should act the same way as they do in file history: display only relevant changes. But since in the Log several paths can be selected in the structure filter, both actions should be altered to support this. This requires addition of a new api method showDiffForPaths to the VcsLogDiffHandler.