external build: fix CompileScope::isAffected(BuildTarget,File) method for targets which need to be fully compiled
If all files from the passed build target are include into the compilation scope, the method should return 'true' even if the scope includes only some files from some other targets. This is needed to properly implement 'Validate JSP' action when JSP validation is performed as part of the external build process (see IDEA-206443). In that case JspValidationTarget takes output of an ArtifactTarget as its input, so in order to perform validation for selected files we create CompileScope which includes these files in ArtifactTarget and also includes the whole validation target.