javaee: rewrite JSP validation to run inside external build process and download required JARs (IDEA-206443)
This fixes several issues when validation wasn't performed correctly because required files weren't copied to the artifact; also the validator don't process PSI inside the IDE process to analyze dependencies (it processes class-files inside the build process), this will help to avoid performance issues.

The validator's JARs are removed from Git repository and don't bundled with the product anymore; they are attached as Maven libraries with 'provided' scope to the sources, and downloaded automatically when validation is invoked for the first time. This also helps to avoid issues with updating libraries on which they depend (e.g. Ant).

Tests are rewritten to use a simple mock implementation of Jasper validator; this way they run faster, don't depend on external libraries and don't require specific JDK version (previously it was needed to use Java 6 to run them).