add isRename method in VirtualFilePropertyEvent
There is a "reparse of file" event which use VirtualFile.PROP_NAME. Many implementations need to handle only rename events but not reparse one. To eliminate possible wrong implementations and code duplications of "isRename" method it is better to add it directly in VirtualFilePropertyEvent.
Note that this API is experimental and can be changed in future.