EA-123951 - TDDE: TraceableDisposable.throwDisposalError
This fixes the problem appeared after the fix for CPP-11610 (commit dc782643).
offset<->logicalPosition conversions cannot be performed for a disposed editor, and we cannot check that editor isn't disposed in a background thread under read action
without race condition, as editor can be closed without write action. So we'll just avoid performing that conversion in the background thread.