[vcs-log] recreate text filter when text filter settings change
Previously, VcsLogClassicFilterUi.getFilters recreated text area filters (but not other filters) each time, but this was changed during filters refactoring. So now text filter explicitly need to be recreated, and it is better to do so inside TextFilterModel and not in VcsLogUiImpl which should not depend on such specifics.

follow-up: 0ed10240d519ca4f14448e34f014bce337feaf61