[git] replace vfs refresh with clm refresh
GitFileHistory asks ChangeListManager for a correct file path in case it was moved but not committed. In `test cyclic rename` occasionally file path before the move gets returned, even though the move is committed. To avoid that CLM should be refreshed before calling GitFileHistory. Separate VFS refresh is however unnecessary.