ConstrainedExecution: Simplify by removing ConstraintSchedulingExecutor
Rename ConstrainedExecutionEx to ConstrainedExecutionScheduler to
emphasize it's an implementation detail, not an extension over the
ConstrainedExecution interface, and remove indirection through
Executor. Although it provided a good level of encapsulation, it was
still complicating the things a bit unnecessarily.

Leave only a single scheduleWithinConstraints() method, which replaces
the previous createConstraintSchedulingExecutor().execute(runnable)
call chain.