platform: register actions and components in proper order
Array returned by PluginManagerCore.getPlugins isn't properly sorted (a plugin may come before its dependency if it comes via a module dependency) so ActionManagerImpl.registerPluginActions may fail when a plugins tries to add an action to a group registered by a plugin it depends on. Reproduced for dependency between UML and Java plugin when extracting Java support to a separate plugin (IDEA-195719).