Instead of pre-populating don't use the cache at all for up-to-date skeletons
Pre-populating has proven to be unreliable because of sub-modules -- some binary
modules produce several entries in sys.modules, and we can't collect them reliably
back in the per-binary cache directory.

An alternative approach is just to keep all the checks on skeletons validity
on Java side. This way generator is not launched altogether and we don't
re-generate anything to fill a missing entry in the cache for existing sdk
skeletons. The downside is that the cache won't fill as quickly as we'd like
to, but we are backed up by prebuilt skeletons for stdlib anyway.

GitOrigin-RevId: 36fd198ac9998d61447ee20fe47a557c84544ab8