actions: use the same InspectCodeGroup action group in IntelliJ IDEA and other IntelliJ-based IDEs
Now 'InspectCodeGroup' is declared in the same way in JavaIdePlugin.xml and PlatformLangActionManager.xml (later we can reused PlatformLangActionManager.xml from JavaIdePlugin.xml to avoid duplication) and IntelliJ IDEA includes it in a separate 'Analyze' menu whereas other IDEs append it to 'Code' menu. This is needed for consistency and to allow extracting Java support to a separate plugin (IDEA-195719).