expose /api/startUpMeasurement
Remove compiled version of start-up-visualizer from source code in favour of externally hosted (for now - https://ij-perf.develar.org)
In general, nothing stops us to continue providing because still it is small text only js file, but not clear does it worth it or not.

Migrate from parcel to webpack because:

1. SRI support — https://github.com/waysact/webpack-subresource-integrity (parcel issue: https://github.com/parcel-bundler/parcel/issues/2003)
2. Full typescript support (including type checking) — https://github.com/TypeStrong/ts-loader & https://github.com/Realytics/fork-ts-checker-webpack-plugin (parcel issue: https://github.com/parcel-bundler/parcel/issues/1378)
3. Externals support — https://webpack.js.org/configuration/externals/ (parcel issue: https://github.com/parcel-bundler/parcel/issues/144) There is workaround, but still because of lack type checking for typescript parcel is no-go.

So, parcel is good for very small projects during experimenting but not suitable for any production usage. Yeach, a lot of things in webpack can be simplified without sacrificing user-friendly (so funny bug that documented flag `optimize-minimize` doesn't work if `mode` also specified - and so on), but as no webpack alternatives at the moment, it is ok to write monstrous webpack.config.js