[vcs-log] select new indexing limit greater than elapsed time
When computer was put to sleep during indexing, reported elapsed time could be incorrect and greater that expected. So it should be taken into account when selecting new indexing limit when user presses Resume. The new code tries to select the first number that is greater than elapsed time and is divisible by the limit value from registry.

IDEA-191688