actions: reimplement logic used to determine whether 'New Scratch File' action is shown in 'New' submenu
Previously it was checked via 'PlatformOpenProjectGroup' action group, but this group may not be added in all IDEs, so the check didn't work properly e.g. for Rider. Now it's checked in a more straightforward and reliable way by inspecting 'New' group.

NewActionGroup.isActionInNewPopupMenu method is needed to do a similar thing for other actions which are located in 'New' submenu in IDEA but will be shown directly in 'File' menu for other IDEs with Java plugin (IDEA-195719).