Check that build view is not null before sending events
There is a NPE when multiple projects are open in Android Studio, the
reason is that if there are multiple build views, the map that holds
them can be updated by other threads. Every time this map::get method is
called in MultipleBuildsView.java the code checks that the value is not
null, except for a case.

This change also makes sure that the view is not null in the missing
part.

This is on PR from IDEA at:
https://github.com/JetBrains/intellij-community/pull/1073

Bug: 120118285
Test: Manually confirm that the NPE exception is no longer seen when
multiple builds are present.

Change-Id: Ib7e771f90b622b509a5037ade141aed06511da74