IDEA-CR-44487: actions: improve fix of presentation of Java-specific New Project/Module action in non-java IDEs (IDEA-195719)
This is an improvement of the previous fix (7108be979f993). Now the naming logic is extracted to a method in NewProjectOrModuleAction interface, and variants for different actions are specified explicitly in ProjectBundle.properties. Also text isn't changed for actions on Welcome Screen.

(cherry picked from commit 5cbc50acdc88857001349b936f4bcf258f3f5246)