IDEA-208453: Fix a memory leak in GitXmlRpcHandlerService
GitXmlRpcHandlerService.registerHandler was registering anonymous Disposable objects with the Disposer. Since the parentDisposable passed to the registerHandler method is always a Project, these anonymous Disposables accumulated in the Disposer's tree for the life time of the project creating an memory leak.

This change replaces anonymous Disposables with instances of the HandlerWrapper class that are disposed either by the unregisterHandler method or together with the project, whichever comes first.