java: don't check superclass to decide whether a class has default constructor
Accordingly to JLS 8.8.9 default constructor is implicitly declared if there are no constructor declarations. If the superclass doesn't have an accessible no-arg constructor, it'll be a compile-time error which shouldn't bother those who tries to use the derived class.

This also fixes incorrect behavior when super-class declares package-private constructor (IDEA-208482).