vcs-log: support range filters defined by ref1..ref2
The range filter displays only commits which are both descendants of ref1 and ancestors of ref2.

With a proper UI such a filter can be useful for:
* comparing two branches in a log tab (IDEA-207634),
* displaying commits exclusive for a branch (IDEA-126079),
* displaying a fixed set of commits between two points (IDEA-52643, IDEA-168060)

The new range filter is defined in the same branch selector popup using the syntax of Git: `..`

Commits from range filter are collected by walking up & down the graph. The Filterer first collects the commits from the range filters, and afterwards filters them by details filters (if any).