PY-34478 Properly follow assignment chains for type aliases
We used to implicitly rely on PyTypingTypeProvider.getClassType() for this purpose
that, in turn, uses normal resolve machinery of a PyReference. But it (rightfully)
didn't work for references inside fragments created from text of alias RHS parts
as we prohibit resolving in them to avoid unstubbing of their containing files.