git: use a more clear name getRepositoryForFile
getRepositoryFor() is not clear enough, since it may mean both "for file" and "for root". Although there is usually no harm in passing a root to a method which accepts a file (although it could affect some submodule root logic), it is still better to follow existing conventions defined in GitRepositoryManager.getRepositoryForRoot/File.