java: make JavaIdePlugin.xml consistent with PlatformLangPlugin.xml
Move registration of extensions and services which are present in PlatformLangPlugin.xml (and therefore included in all small IDEs) to the beginning of JavaIdePlugin.xml. Those which have special implementation in IntelliJ IDEA are now explicitly override the default implementation.

This is a preliminary step to reuse PlatformLangPlugin.xml in IntelliJ IDEA as well to make all IDEs consistent and avoid duplication. This is also needed to properly extract Java support to a separate plugin (IDEA-195719).