Track strings and primitive arrays in DebugReflectionUtil
These are currently considered "trivial", meaning they can't hold
interesting references to other objects. However, this means BLeak can't
record them in the graph, and consequently cannot include their sizes
when it estimates leak severity. Strings and primitive arrays easily
account for over half the data in the heap, so this is an important
omission.

This makes the resulting graph approximately 30% larger.

Fixes: 127320441
Change-Id: I34c9d0ff479ee1887b53b390b4228e9362415bf1