Maven: Import folders from multiple executions
With maven-build-helper, one can define multiple executions that can
each add source or resource folders to the project. With the current
implementation, IntelliJ only considers the very first execution that
matches the given maven goal, like add-sources. However, it is perfectly
fine with Maven to define multiple executions with the same goal. This
also happens automatically, for example when multiple Maven profiles are
defined and active, that each call add-source of the maven-build-helper.

With this commit, I changed the folder importing strategy, so that not
only the first matching execution is considered, but all executions that
call the appropriate Maven goal.

I also added a test case that fails on current master, but passes with
this commit. It simulates a project structure with multiple additional
sources and resources folders, with each of them added in different
executions of the maven-build-helper. This can be a common scenario if
one extensively uses Maven profiles to trigger specific actions.