Register `esc` action for AbstractPopup only if myCancelKey is enabled
The previous solution for closing heavyweight hints caused VIM-1589 issue (`esc` action was not fired
while autocompletion popup is open). This is also a root cause for IDEA-203019 and 0951ab6a as a workaround.

Extracting `if (myCancelKeyEnabled)` out of the action handler has the same functionality, but passes `esc` action
to further handlers. This solves VIM-1589 and IDEA-203019 (and 0951ab6a is no more needed).