java: register the common PlatformProjectCheckoutListener in Java-based IDEs and disable it for them
Currently all IDEs except IntelliJ IDEA include PlatformProjectCheckoutListener (registered in PlatformLangPlugin.xml) which suggest to open the checked out directory as a project. In order to reuse PlatformLangPlugin.xml in IntelliJ IDEA and extract Java support as a separate plugin (IDEA-195719) we need to register PlatformProjectCheckoutListener in IntelliJ IDEA as well and disable it there.