idea: reuse PlatformLangComponents.xml instead of using custom IdeaComponents.xml in IDEA
Components which were present in IdeaComponents and not present in PlatformLangComponents are now registered in JavaIdePlugin directly. PlatformVcsDetector is the only component which is included into PlatformLangComponents but wasn't included into IdeaComponents, but adding it to IntelliJ IDEA shouldn't do any harm.

This is a preliminary step to reuse PlatformLangPlugin.xml in IntelliJ IDEA as well to make all IDEs consistent and avoid duplication. This is also needed to properly extract Java support to a separate plugin (IDEA-195719).