idea: reuse PlatformLangActionManager.xml instead of copying its content
Groups which include analysis actions into Code, Tools and popup menus are now overriden with empty groups in JavaActions.xml so these actions will be shown in a separate Analyze menu as before if Java support is enabled.

This is a preliminary step to reuse PlatformLangPlugin.xml in IntelliJ IDEA as well to make all IDEs consistent and avoid duplication. This is also needed to properly extract Java support to a separate plugin (IDEA-195719).