Remove unused `esc` listener
The problem is that additional listener (that was removed with this commit)
overrides existing esc listeners of content.

Additional listener for esc is redundant because it's never get called.
`com.intellij.ui.popup.AbstractPopup#dispatchKeyEvent` prevents listeners
execution if it returns true. So, if `myCancelKeyEnabled` is `false`,
listener is not executed because of `if` condition and in case of `true`
listeners are not executed.

As long as this additional listener doesn't override existing listeners,
`hide()` function is called with `myEscListener`.

Review: IDEA-CR-45219

(cherry picked from commit f175fc8aba8cc83a2fc97f7b7dd732c000362bd0)