platform: use WebModuleTypeBase instead of specific implementations
'WEB_MODULE' resolves to WebModuleType in IntelliJ IDEA and to PlatformWebModuleType in WebStorm and PhpStorm, so WebModuleType.getInstance will fail with ClassCastException if it's invoked in WebStorm/PhpStorm and PlatformWebModuleType.getInstance will fail if it's invoked in IDEA. Now the both methods are deprecated and PlatformWebModuleType.getInstance which work in all IDEs is used instead. This is needed to use Java support as a plugin in WebStorm (IDEA-195719).

(cherry picked from commit 777ba26d5c4f286d6282076c35643792531b4494, IDEA-CR-45436)