IDEA-208611 Fix a memory leak in IdeGlassPaneImpl._addListener
This change introduces a new class DisposableWrapperList that supports associating elements with disposables, so that the elements are removed when their disposables are disposed.

I didn't attempt to fix the leak in IdeGlassPaneImpl.addPainter or any leaks outside of IdeGlassPanelImpl and EventDispatcher to keep the size of the change manageable.