register top-level BuildMenu action group in platform
Both Java support and CLion adds their actions to Build menu in the main menu, so in order to properly support CLion with Java plugin installed (IDEA-195719) the group registration is moved to the platform. The menu won't be visiable in IDEs which don't add their actions to the Build menu.