IDEA project configuration: enable inspection about access to package-private members from different module
This is needed to avoid such code because it'll lead to IllegalAccessError if e.g. Java support code is loaded as a plugin via a different classloader (see IDEA-200277). Modules belonging to same plugins are listed in the inspection settings and accessing package-private member between classes from them aren't reported.

There are some false-positives because of bugs in the inspection (see IDEA-210216, IDEA-210253), they are temporarily suppressed.