PY-35141 Properly discover conda executable relative to python.exe on Windows
Previously we mistakenly searched for "Scripts" directory, where conda.exe
normally resides, in the catalog one level above the root of a conda environment.
It went unnoticed for a long time as when we didn't succeed in finding conda.exe
relatively to a Python executable we scan through the list of known default
installation paths of Anaconda and often eventually locate either the right one
or some other conda.exe that is apparently still sufficient for activation.

This problem also showed up on our TC agents where there are no existing
Anaconda installations and new ones are installed in a non-standard location.
Thus, environments couldn't be activated anyhow and generation of pre-built
skeletons failed for most of modules of Anaconda 3 (PY-34604).